+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.0 (under development)
 This is the initial version setting up the base system
 
+- `CNG` Renamed message event type: Published -> Created
+- `NEW` Add publishedAt field. Timestamp in UTC
 - `NEW` Add geocode field. Expects a geometry polygon as WKT 
 - `NEW` Basic message event and payload model
 - `NEW` Add message event deserializer